feat: 공통 Error 컴포넌트 작업
