+++ src/components/ui/spinner.tsx
@@ -1,4 +1,5 @@
 import { cva, type VariantProps } from 'class-variance-authority';
+import { PropsWithChildren } from 'react';
 
 import { cn } from '@/lib/utils';
